requirements.txt
gradio==5.49.1
python-dotenv==1.0.1
google-generativeai==0.8.3
langchain==0.3.10
langchain-google-genai==2.0.7
langgraph==0.2.60
toon-format>=0.9.0b1
pydantic>=2.9.2
aiohttp
orjson
